requirements.txt
google-api-python-client
google-auth
google-auth-httplib2
beautifulsoup4
icalendar
pytz
pytest-mock
python-dotenv
